Update .gitignore, README, and configuration files; remove brand classifier agent
- Added entries to .gitignore for temporary files, model files, and cache directories.
- Expanded README to provide a detailed overview of the RiseApp architecture and processes.
- Updated config.yaml to allow for local model usage and removed references to the brand classifier agent.
- Deleted brand classifier agent files and prompts to streamline the project focus on lineage detection.
- Adjusted embedding factory to support local model paths and updated tests accordingly.

+++ src/embedding/huggingface_embedding.py
@@ -24,7 +24,7 @@ class HuggingFaceEmbedding(BaseEmbedding):
     def embed(self, texts: List[str]) -> np.ndarray:
         if not texts:
             return np.zeros((0, self.embedding_dim()), dtype=np.float32)
-        out = self._model.encode(texts, convert_to_numpy=True, dtype="float32")
+        out = self._model.encode(texts, convert_to_numpy=True)
         return np.asarray(out, dtype=np.float32)
 
     def embedding_dim(self) -> int:
